+++ scripts/parameter/main.js
@@ -26,7 +26,7 @@ function main() {
   // save input fields
   const fields = document.querySelectorAll(".field,.codecompanyfield,.codecustomerfield,.parameternamefield,.newvaluefield,.changereasonfield");
   for (const field of fields) {
-    field.addEventListener("keyup", fieldChange);
+    field.addEventListener("input", fieldChange);
   }
 
   // update hover-overs on load
@@ -34,12 +34,6 @@ function main() {
   for (const field of allFields) {
     field.title = field.value;
   }
-
-  // show parameter search options
-  // const paramFields = document.querySelectorAll(".parameternamefield");
-  // for (const field of paramFields) {
-  //   field.addEventListener("keydown",parameterOptionsShow);
-  // }
 
   // keydown actions inside grid
   const lineFields = document.querySelectorAll(".codecompanyfield,.codecustomerfield,.parameternamefield,.parameteroptionsfield,.previousvaluefield,.newvaluefield,.changereasonfield");
@@ -92,9 +86,9 @@ function fieldChange(event) {
   field.title = field.value;
 
   // update field outline and tooltip
-  if (['codecompanyfield','codecustomerfield','parameternamefield'].includes(fieldType)) {
+  if (['codecompanyfield','codecustomerfield','parameternamefield','parameteroptionsfield'].includes(fieldType)) {
     // codecompany, codecustomeror parametername: check all three in all rows
-    const rowFields = document.querySelectorAll(".codecompanyfield,.codecustomerfield,.parameternamefield");
+    const rowFields = document.querySelectorAll(".codecompanyfield,.codecustomerfield,.parameternamefield,.parameteroptionsfield");
     for (const rowField of rowFields) {
       updateFieldOutlineAndTooltip(rowField.id);
     }
@@ -112,12 +106,32 @@ function fieldChange(event) {
 
   // additional actions
   switch (fieldType) {
+    case 'codecompanyfield':
+    case 'codecustomerfield':
+    case 'parameternamefield':
+      clearGetResponse(field.id);
+      clearSetResponse(field.id);
+      clearPreviousValue(field.id);
+      break;
     case 'previous':
       updateHighlightSet();
+    case 'newvaluefield':
       updateCurrentValuesHighlight();
+      clearSetResponse(field.id);
       break;
 
     case 'environment':
+      // clear responses, previous values
+      const nofLines = parseInt(document.getElementById('noflines').value);
+
+      for (let index = 0; index<nofLines; index++) {
+        var fieldId = "codecompany" + index.toString();
+        clearGetResponse(fieldId);
+        clearSetResponse(fieldId);
+        clearPreviousValue(fieldId);
+      }
+
+      // update panel coloring
       showProd();
       break;
 
@@ -135,6 +149,41 @@ function fieldChange(event) {
       document.getElementById("savename").innerHTML = nameFromPath(field.value);
       break;
   }
+}
+
+function clearValue(id, index) {
+  const text = id + '|' + index;
+  vscodeApi.postMessage({ command: "clearvalue", text: text });
+}
+
+function clearGetResponse(fieldId) {
+  const field = document.getElementById(fieldId);
+  const index = field.getAttribute("index");
+  const getResponse = document.getElementById("getresponsefieldoption" + index);
+
+  getResponse.innerHTML = "";
+  getResponse.title = "";
+  clearValue('getresponse',index);
+}
+
+function clearSetResponse(fieldId) {
+  const field = document.getElementById(fieldId);
+  const index = field.getAttribute("index");
+  const setResponse = document.getElementById("setresponsefieldoption" + index);
+
+  setResponse.innerHTML = "";
+  setResponse.title = "";
+  clearValue('setresponse',index);
+}
+
+function clearPreviousValue(fieldId) {
+  const field = document.getElementById(fieldId);
+  const index = field.getAttribute("index");
+  const previousField = document.getElementById("previousvalue" + index);
+
+  previousField.value = "";
+  previousField.title = "";
+  clearValue('previousvalue',index);
 }
 
 function gridKeydown(event) {
@@ -313,7 +362,7 @@ function updateParameterFromOptions(event) {
 
   // update parameter value
   parameterField.value = value;
-  parameterField.dispatchEvent(new Event('keyup'));
+  parameterField.dispatchEvent(new InputEvent('input'));
 }
 
 function checkIfDuplicate(row) {
@@ -351,6 +400,7 @@ function getCompanyName(codeCompany) {
 }
 
 function showProd() {
+  // update panel coloring
   const field = document.getElementById("environment");
   const documentBgColor = getComputedStyle(field).getPropertyValue('--vscode-editor-background');
   const fieldBgColor = getComputedStyle(field).getPropertyValue('--vscode-dropdown-background');
@@ -394,7 +444,7 @@ function processingGet(push = false) {
   const isProcessing = !document.getElementById("processingget").hidden;
 
   if (isProcessing) {
-    const disableFields = document.querySelectorAll("#environment,#noflines,#previous,#setparameters,#getparameters,#load,.codecompanyfield,.codecustomerfield,.parameternamefield");
+    const disableFields = document.querySelectorAll("#environment,#noflines,#previous,#setparameters,#getparameters,#load,.codecompanyfield,.codecustomerfield,.parameternamefield,.parameteroptionsfield");
     for (const dField of disableFields) {
       dField.disabled = true;
     }
@@ -408,7 +458,7 @@ function processingSet(push = false) {
   const isProcessing = !document.getElementById("processingset").hidden;
 
   if (isProcessing) {
-    const disableFields = document.querySelectorAll("#environment,#noflines,#previous,#allchangereasons,#setparameters,#getparameters,#load,.codecompanyfield,.codecustomerfield,.parameternamefield,.previousvaluefield,.newvaluefield,.changereasonfield");
+    const disableFields = document.querySelectorAll("#environment,#noflines,#previous,#allchangereasons,#setparameters,#getparameters,#load,.codecompanyfield,.codecustomerfield,.parameternamefield,.parameteroptionsfield,.previousvaluefield,.newvaluefield,.changereasonfield");
     for (const dField of disableFields) {
       dField.disabled = true;
     }
@@ -515,7 +565,7 @@ function updateFieldOutlineAndTooltip(fieldId) {
     updateEmpty(field.id);
   } else if (check !== '' && check !== null) {
     updateWrong(field.id, getContentHint(fieldType));
-  } else if (['codecompanyfield','codecustomerfield','parameternamefield'].includes(fieldType) && checkIfDuplicate(+field.getAttribute('index'))) {
+  } else if (['codecompanyfield','codecustomerfield','parameternamefield','parameteroptionsfield'].includes(fieldType) && checkIfDuplicate(+field.getAttribute('index'))) {
     updateWrong(field.id, 'Company/customer/parameter combination is duplicate');
   } else {
     updateRight(field.id);